adds error handling for dao

+++ src/main/java/com/fabian/repetitionEmployeeAPI/dao/EmployeeDAOImpl.java
@@ -1,5 +1,6 @@
 package com.fabian.repetitionEmployeeAPI.dao;
 
+import java.io.UncheckedIOException;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,7 +11,7 @@ import com.fabian.repetitionEmployeeAPI.component.Employee;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.Query;
 import jakarta.persistence.TypedQuery;
-import jakarta.transaction.Transactional;
+import org.springframework.transaction.annotation.Transactional;
 
 @Repository
 public class EmployeeDAOImpl implements EmployeeDAO {
@@ -24,26 +25,23 @@ public class EmployeeDAOImpl implements EmployeeDAO {
 	
 	@Override
 	@Transactional
-	public void saveEmployee(Employee employee) throws Exception {
-		try {
+	public Employee saveEmployee(Employee employee) {
+		  try {
 			if (employee.getId() != 0) {
-		        entityManager.merge(employee); // Use merge() for updates
+		        return entityManager.merge(employee); 
 		    } else {
-		        entityManager.persist(employee);
-		    }
+		         entityManager.persist(employee);
+		        return employee; 
+		    } 
 		}
-		
 		catch(Exception e) {
-			throw new Exception(e.getMessage());
+			throw new DatabaseException("Database issue with saving/updating employee");
 		}
 	}
 
 	
 	@Override
-	public Employee getEmployeeById(int id) throws Exception {
-		if(id < 0) {
-			//error handling here
-		}
+	public Employee getEmployeeById(int id) {
 		TypedQuery<Employee> theQuery = entityManager.createQuery("from Employee where id=:uid", Employee.class);
 		theQuery.setParameter("uid", id);
 
@@ -51,7 +49,10 @@ public class EmployeeDAOImpl implements EmployeeDAO {
 		try {
 			employee = theQuery.getSingleResult();
 		} catch (Exception e) {
-			throw new Exception(e.getMessage());	
+			if(employee == null) {
+				throw new EmployeeNotFoundException("Requested Employee could not be found");
+			}
+			throw new DatabaseException("Database issue with retrieving employee");
 		}
 		return employee;
 	}
@@ -59,25 +60,47 @@ public class EmployeeDAOImpl implements EmployeeDAO {
 
 
 	@Override
-	public List<Employee> getAllEmployees() throws Exception {
+	public List<Employee> getAllEmployees() {
 		TypedQuery<Employee> theQuery = entityManager.createQuery("from Employee", Employee.class);
 		List<Employee> employees = null;
 		try {
-			employees = theQuery.getResultList();
-		} catch (Exception e) {
-			throw new Exception(e.getMessage());	
+		employees = theQuery.getResultList();
+		}
+		catch(Exception exc) {
+		throw new DatabaseException("There is a database issue with retrieving the entries");	
+		}
+		if(employees.size() == 0) {
+		throw new EmployeeNotFoundException("Requested Employees could not be found");
 		}
 		return employees;
 	}
 
-
-
 	@Override
 	@Transactional
-	public void deleteEmployeeById(int id) throws Exception {
-		Query query = entityManager.createQuery("DELETE FROM Employee e WHERE e.id = :id");
-		query.setParameter("id", id);
-		query.executeUpdate();
+	public String deleteEmployeeById(int id) {
+		
+	Employee employee = entityManager.find(Employee.class, id);
+		
+	  if (employee == null) {
+            throw new EmployeeNotFoundException("Employee with ID " + id + " does not exist.");
+        }
+		
+		try {
+			Query query = entityManager.createQuery("DELETE FROM Employee e WHERE e.id = :id");
+			query.setParameter("id", id);
+			int numbOfEntities = query.executeUpdate(); 
+			if (numbOfEntities == 0) {
+	            throw new DatabaseException("Failed to delete employee with ID " + id);
+			}
+	
+			return "Employee with id" + id + " deleted";
+		}
+		
+		catch (Exception exc) {
+			throw new DatabaseException("There is an issue with the database");
+		}
+		
+		
 	}
 
 }
